BIOS-592: weblib.sh : initial support for common parsing of HTTP headers in curl stderr output (currently only warns by default until tests are converted to match)

+++ tests/CI/weblib.sh
@@ -42,6 +42,19 @@ echo "INFO-WEBLIB: Will use BASE_URL = '$BASE_URL'"
 # Should the test suite abort if "curl" errors out?
 [ -z "$WEBLIB_CURLFAIL" ] && WEBLIB_CURLFAIL=yes
 
+# Should the test suite abort if "curl" sees HTTP error codes?
+# This can be overridden on a per-call basis for those api_get's
+# where we do expect errors as the proper response.
+#       fatal   Fail if errors are detected in stderr contents
+#       ignore  Don't care, and don't test
+#       expect  Fail if stderr is not empty but result is OK
+#       *       Anything else gives a warning if error was matched and goes on
+[ -z "$WEBLIB_CURLFAIL_HTTPERRORS" ] && WEBLIB_CURLFAIL_HTTPERRORS=warn
+[ -z "$WEBLIB_HTTPERRORS_REGEX" ] && WEBLIB_HTTPERRORS_REGEX='HTTP/[^ ]+ [45]'
+
+# Print out the CURL stdout and stderr (via FD#3)?
+[ -z "$WEBLIB_TRACE_CURL" ] && WEBLIB_TRACE_CURL=no
+
 [ -z "$CHECKOUTDIR" ] && CHECKOUTDIR="`dirname $0`/../.."
 [ -z "$JSONSH" ] && JSONSH="$CHECKOUTDIR/tools/JSON.sh"
 
@@ -108,9 +121,12 @@ test_it() {
 
 ### This is what we will sig-kill if needed
 _PID_TESTER=$$
+RES_CURL=0
 trap_break() {
     ### This SIGUSR1 handler is reserved for CURL failures
-    echo "CI-WEBLIB-ERROR-WEB: curl program failed, aborting test suite" >&2
+    [ "$RES_CURL" = 0 ] && \
+        echo "CI-WEBLIB-ERROR-WEB: curl program failed, aborting test suite" >&2 || \
+        echo "CI-WEBLIB-ERROR-WEB: curl program failed ($RES_CURL), aborting test suite" >&2
     echo "CI-WEBLIB-FATAL-BREAK: Got forced interruption signal" >&2
     WEBLIB_FORCEABORT=yes
 
@@ -121,28 +137,75 @@ trap_break() {
 trap "trap_break" SIGUSR1
 
 CURL() {
-    curl "$@"
-    RES_CURL=$?
+    _TMP_CURL="/tmp/.bios_weblib_curl.$$"
+    /bin/rm -f "$_TMP_CURL" 2>/dev/null
+    if touch "$_TMP_CURL" && chmod 600 "$_TMP_CURL" ; then
+        OUT_CURL="`curl --stderr "$_TMP_CURL" "$@"`"
+        RES_CURL=$?
+        ERR_CURL="`cat "$_TMP_CURL"`"
+        /bin/rm -f "$_TMP_CURL" 2>/dev/null
+        echo "$ERR_CURL" >&2
+        echo "$OUT_CURL"
+    else
+        OUT_CURL=""
+        ERR_CURL=""
+#        OUT_CURL="`curl "$@"`"
+        curl "$@"
+        RES_CURL=$?
+#        echo "$OUT_CURL"
+    fi
+
     if [ $RES_CURL != 0 ]; then
         ### Based on caller redirections, this output may never be seen
         echo "CI-WEBLIB-ERROR-CURL: 'curl $@' program failed ($RES_CURL)," \
-            "perhaps the web server is not available or has crashed?" >&2
+            "perhaps the web server is not available or has crashed?" >&3
+    fi
 
-        if [ x"$WEBLIB_CURLFAIL" = xyes ]; then
-            kill -SIGUSR1 $_PID_TESTER $$ >/dev/null 2>&1
-            # exit $RES_CURL
+    if [ -n "$ERR_CURL" -a x"$WEBLIB_CURLFAIL_HTTPERRORS" != xignore ]; then
+        ERR_MATCH="`( echo "$ERR_CURL"; echo "" ) | tr '\r' '\n' | egrep '^ *< '"$WEBLIB_HTTPERRORS_REGEX"`"
+        if [ -n "$ERR_MATCH" ]; then
+            echo "CI-WEBLIB-WARN-CURL: Last response headers matched an HTTP" \
+                "error code: '$ERR_MATCH'" >&3
+            if [ x"$WEBLIB_CURLFAIL_HTTPERRORS" = xfatal ]; then
+                echo "CI-WEBLIB-ERROR-CURL: WEBLIB_CURLFAIL_HTTPERRORS=fatal" \
+                    "is set, so failing now" >&3
+                RES_CURL=123
+            fi
+        else
+            ### No match
+            if [ x"$WEBLIB_CURLFAIL_HTTPERRORS" = xexpect ]; then
+                echo "CI-WEBLIB-ERROR-CURL: Last response headers did not match" \
+                    "an HTTP error pattern '$WEBLIB_HTTPERRORS_REGEX', while" \
+                    "WEBLIB_CURLFAIL_HTTPERRORS=expect is set, so failing now" >&3
+                RES_CURL=124
+            fi
         fi
+    fi
+
+    if [ x"$WEBLIB_TRACE_CURL" = xyes -o $RES_CURL != 0 ]; then
+        echo "CI-WEBLIB-TRACE-CURL: curl $@: RES=$RES"
+        echo "=== ERR vvv"
+        echo "$ERR_CURL"
+        echo "/// ERR ^^^"
+        echo "=== OUT vvv"
+        echo "$OUT_CURL"
+        echo "/// OUT ^^^"
+    fi >&3
+
+    if [ $RES_CURL != 0 -a x"$WEBLIB_CURLFAIL" = xyes ]; then
+        kill -SIGUSR1 $_PID_TESTER $$ >/dev/null 2>&1
+        # exit $RES_CURL
     fi
 
     return $RES_CURL
 }
 
 api_get() {
-    CURL --insecure -v --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -v --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 api_get_content() {
-    CURL --insecure "$BASE_URL$1" 2>/dev/null
+    CURL --insecure "$BASE_URL$1" 3>&2 2>/dev/null
 }
 
 api_get_json() {
@@ -152,13 +215,13 @@ api_get_json() {
         return $?
     fi
 
-    CURL --insecure -v --progress-bar "$BASE_URL$1" 2> /dev/null \
+    CURL --insecure -v --progress-bar "$BASE_URL$1" 3>&2 2> /dev/null \
     | $JSONSH -N
 }
 
 api_get_json_sed() {
     ### Old approach to strip any whitespace including linebreaks from JSON
-    CURL --insecure -v --progress-bar "$BASE_URL$1" 2> /dev/null \
+    CURL --insecure -v --progress-bar "$BASE_URL$1" 3>&2 2> /dev/null \
     | tr \\n \  | sed -e 's|[[:blank:]]\+||g' -e 's|$|\n|'
 }
 
@@ -169,7 +232,7 @@ api_get_jsonv() {
 }
 
 api_post() {
-    CURL --insecure -v -d "$2" --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -v -d "$2" --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 ### Flag for _api_get_token_certainPlus()
@@ -248,7 +311,8 @@ api_auth_post() {
     #	$1	Relative URL for API call
     #	$2	POST data
     TOKEN="`_api_get_token`"
-    CURL --insecure -v -H "Authorization: Bearer $TOKEN" -d "$2" --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -H "Authorization: Bearer $TOKEN" -d "$2" \
+        -v --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 api_auth_post_content() {
@@ -256,7 +320,8 @@ api_auth_post_content() {
     #	$1	Relative URL for API call
     #	$2	POST data
     TOKEN="`_api_get_token`"
-    CURL --insecure -H "Authorization: Bearer $TOKEN" -d "$2" "$BASE_URL$1" 2>/dev/null
+    CURL --insecure -H "Authorization: Bearer $TOKEN" -d "$2" \
+        "$BASE_URL$1" 3>&2 2>/dev/null
 }
 
 api_auth_post_wToken() {
@@ -264,7 +329,8 @@ api_auth_post_wToken() {
     #	$1	Relative URL for API call
     #	$2	POST data
     TOKEN="`_api_get_token`"
-    CURL --insecure -v -d "access_token=$TOKEN&$2" --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -d "access_token=$TOKEN&$2" \
+        -v --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 api_auth_post_content_wToken() {
@@ -272,12 +338,13 @@ api_auth_post_content_wToken() {
     #	$1	Relative URL for API call
     #	$2	POST data
     TOKEN="`_api_get_token`"
-    CURL --insecure -d "access_token=$TOKEN&$2" "$BASE_URL$1" 2>/dev/null
+    CURL --insecure -d "access_token=$TOKEN&$2" "$BASE_URL$1" 3>&2 2>/dev/null
 }
 
 api_auth_delete() {
     TOKEN="`_api_get_token`"
-    CURL --insecure -v -H "Authorization: Bearer $TOKEN" -X "DELETE" --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -H "Authorization: Bearer $TOKEN" -X "DELETE" \
+        -v --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 api_auth_put() {
@@ -285,12 +352,14 @@ api_auth_put() {
     #	$1	Relative URL for API call
     #	$2	PUT data
     TOKEN="`_api_get_token`"
-    CURL --insecure -v -H "Authorization: Bearer $TOKEN" -d "$2" -X "PUT" --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -H "Authorization: Bearer $TOKEN" -d "$2" -X "PUT" \
+        -v --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 api_auth_get() {
     TOKEN="`_api_get_token`"
-    CURL --insecure -v -H "Authorization: Bearer $TOKEN" --progress-bar "$BASE_URL$1" 2>&1
+    CURL --insecure -H "Authorization: Bearer $TOKEN" \
+        -v --progress-bar "$BASE_URL$1" 3>&2 2>&1
 }
 
 api_auth_get_wToken() {
@@ -299,12 +368,13 @@ api_auth_get_wToken() {
     case "$1" in
         *"?"*) URLSEP='&' ;;
     esac
-    CURL --insecure -v --progress-bar "$BASE_URL$1$URLSEP""access_token=$TOKEN" 2>&1
+    CURL --insecure -v --progress-bar \
+        "$BASE_URL$1$URLSEP""access_token=$TOKEN" 3>&2 2>&1
 }
 
 api_auth_get_content() {
     TOKEN="`_api_get_token`"
-    CURL --insecure -H "Authorization: Bearer $TOKEN" "$BASE_URL$1" 2>/dev/null
+    CURL --insecure -H "Authorization: Bearer $TOKEN" "$BASE_URL$1" 3>&2 2>/dev/null
 }
 
 api_auth_get_content_wToken() {
@@ -313,23 +383,25 @@ api_auth_get_content_wToken() {
     case "$1" in
         *"?"*) URLSEP='&' ;;
     esac
-    CURL --insecure "$BASE_URL$1$URLSEP""access_token=$TOKEN" 2>/dev/null
+    CURL --insecure "$BASE_URL$1$URLSEP""access_token=$TOKEN" 3>&2 2>/dev/null
 }
 
 api_auth_get_json() {
     TOKEN="`_api_get_token`"
-    CURL --insecure -v --progress-bar -H "Authorization: Bearer $TOKEN" "$BASE_URL$1" 2> /dev/null \
+    CURL --insecure -v --progress-bar -H "Authorization: Bearer $TOKEN" \
+        "$BASE_URL$1" 3>&2 2> /dev/null \
     | tr \\n \  | sed -e 's|[[:blank:]]\+||g' -e 's|$|\n|'
 }
 
 api_auth_get_jsonv() {
     TOKEN="`_api_get_token`"
-    api_auth_get_json "$@" | python -c "import sys, json; s=sys.stdin.read(); json.loads(s); print(s)"
+    api_auth_get_json "$@" | \
+        python -c "import sys, json; s=sys.stdin.read(); json.loads(s); print(s)"
 }
 
 api_post_json_cmp() {
 #    set -x
-    text=$(CURL --insecure -v --progress-bar -d "$2" "$BASE_URL$1" 2>&1)
+    text=$(CURL --insecure -v --progress-bar -d "$2" "$BASE_URL$1" 3>&2 2>&1)
     res=$(echo "${text}" | egrep "$3")
     if [ -z "$res" ]; then
         return 1
